requirements.txt
flask
jinja2
pdfkit
python-docx
requests
pymupdf
pytesseract
Pillow
markdown2
docx2txt
